+++ pyengine/Systems/__init__.py
@@ -2,5 +2,6 @@ from pyengine.Systems.EntitySystem import EntitySystem
 from pyengine.Systems.MusicSystem import MusicSystem
 from pyengine.Systems.UISystem import UISystem
 from pyengine.Systems.SoundSystem import SoundSystem
+from pyengine.Systems.CameraSystem import CameraSystem
 
-__all__ = ["EntitySystem", "MusicSystem", "UISystem", "SoundSystem"]
+__all__ = ["EntitySystem", "MusicSystem", "UISystem", "SoundSystem", "CameraSystem"]